Update BudwormFir_Simulation.R
This initial simulation of a budworm population explosion succeeds as a proof-of-concept. There are some dynamics that fall short however. Without addressing these (see comments in code), it will be challenging to continue on to the next step of highlighting the system dynamics associated with regular pesticide treatment of the forest.

NEXT STEPS:
 - Ensure fir population is the initial cause of the budworm population explosion
 - Confirm that predator reproduction rates fail to contain the budworm
 - Confirm that fir die-off is the final cause in the ending of the budworm population explosion
 - Introduce pesticide spraying
 - Try to observe the the resultant condition of a near-population-explosion caused by the pesticide as described in the example (Thinking In Systems, page 92)